requirements.txt
binance-connector==3.6.0
numpy==1.26.4
pandas==2.2.2
pyarrow==16.1.0
pydantic==2.8.2
PyYAML==6.0.2
scikit-learn==1.4.2
joblib==1.4.2
rich==13.7.1
matplotlib==3.8.3
ruff==0.6.4
black==24.4.2
pytest==8.2.0
